requirements.txt
discord.py
python-dateutil
